long ==> int

+++ StRoot/Sti/StiTrack.h
@@ -126,8 +126,8 @@ public:
   /// Get chi2 of this track
   virtual double  getChi2()    const=0;
   virtual double  getChi2Max() const=0;   // maximal chi2 
-  virtual void    setFlag(long v)=0;
-  virtual long    getFlag() const=0;
+  virtual void    setFlag(int v)=0;
+  virtual int     getFlag() const=0;
   virtual vector<StiHit*> getHits()=0;
   // Convenience Accessor using a switch
   virtual double  getValue(int key) const;
